+++ spec/PHPCR/Shell/Console/Application/ShellApplicationSpec.php
@@ -5,6 +5,7 @@ namespace spec\PHPCR\Shell\Console\Application;
 use PhpSpec\ObjectBehavior;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 use PHPCR\Shell\Console\Application\EmbeddedApplication;
+use PHPCR\Shell\PhpcrShell;
 
 class ShellApplicationSpec extends ObjectBehavior
 {
@@ -12,7 +13,7 @@ class ShellApplicationSpec extends ObjectBehavior
         ContainerInterface $container
     )
     {
-        $this->beConstructedWith($container, EmbeddedApplication::MODE_COMMAND);
+        $this->beConstructedWith($container, PhpcrShell::MODE_EMBEDDED_COMMAND);
     }
 
     public function it_is_initializable()